agent/requirements.txt
# AI对话框架
chakpy>=0.2.1

# 环境变量管理
python-dotenv>=1.0.0

# 终端美化
rich>=13.0.0
